feat: invitation page sign in or launchpad
